+++ contracts/src/v0.8/ccip/test/helpers/CommitStoreHelper.sol
@@ -5,8 +5,8 @@ import "../../commitStore/CommitStore.sol";
 
 contract CommitStoreHelper is CommitStore {
   constructor(
-    uint256 chainId,
-    uint256 sourceChainId,
+    uint64 chainId,
+    uint64 sourceChainId,
     AFNInterface afn,
     CommitStoreConfig memory config
   ) CommitStore(chainId, sourceChainId, afn, config) {}
